+++ packages/blueconfig/lib/core.js
@@ -9,25 +9,50 @@ const CUSTOMISE_FAILED = cvtError.CUSTOMISE_FAILED
 
 
 /**
- * @return Returns a ConfigObjectModel (= COM, Like DOM but not for Document, for Config)
+ * Cutomize Blueconfig before make a ConfigObjectModel (= COM, Like DOM but not for Document, for Config)
+ *
+ * @example
+ * const Blueconfig = require('blueconfig/core.js')
+ * const blueconfig = new Blueconfig()
+ * console.log(blueconfig.getGettersOrder()) // === [value', 'force']
  *
  * @class
  */
-const Blueconfig = function() {
+const BlueconfigCore = function() {
   this.initPerformer()
 }
 
-module.exports = Blueconfig
+module.exports = BlueconfigCore
 
 
-Blueconfig.prototype.initPerformer = function() {
+/**
+ * Create instance of worker (Parser, Getter, Ruler)
+ *
+ * @return   {this}
+ */
+BlueconfigCore.prototype.initPerformer = function() {
   this.Parser = new ParserInterface()
   this.Getter = new GetterInterface()
   this.Ruler = new RulerInterface()
+
+  return this
 }
 
 
-Blueconfig.prototype.init = function(rawSchema, options) {
+/**
+ * Gets array with getter name in the current order of priority
+ *
+ * @example
+ * const Blueconfig = require('blueconfig/core.js')
+ * const blueconfig = new Blueconfig()
+ * blueconfig.init(rawSchema, options)
+ *
+ * @param   {string|object}   rawSchema    Schema object or a path to a schema JSON file
+ * @param   {object}   [options]    Options. [See ConfigObjectModel constructor doc](./ConfigObjectModel.html)
+ *
+ * @return    {string[]}    Returns current getter order
+ */
+BlueconfigCore.prototype.init = function(rawSchema, options) {
   return new ConfigObjectModel(rawSchema, options, {
     Parser: this.Parser,
     Getter: this.Getter,
@@ -40,11 +65,11 @@ Blueconfig.prototype.init = function(rawSchema, options) {
  * Gets array with getter name in the current order of priority
  *
  * @example
- * blueconfig.getGettersOrder(); // ['default', 'value', 'env', 'arg', 'force']
+ * blueconfig.getGettersOrder() // ['default', 'value', 'env', 'arg', 'force']
  *
  * @return    {string[]}    Returns current getter order
  */
-Blueconfig.prototype.getGettersOrder = function() {
+BlueconfigCore.prototype.getGettersOrder = function() {
   return [...this.Getter.storage.order]
 }
 
@@ -56,21 +81,21 @@ Blueconfig.prototype.getGettersOrder = function() {
  * @see ConfigObjectModel.refreshGetters
  *
  * @example
- * blueconfig.getGettersOrder();
+ * blueconfig.getGettersOrder()
  * // ['default', 'value', 'env', 'arg', 'force']
  *
  * // two ways to do:
- * blueconfig.sortGetters(['default', 'value', 'arg', 'env', 'force']);
- * blueconfig.sortGetters(['default', 'value', 'arg', 'env']); // force is optional and must be the last one
+ * blueconfig.sortGetters(['default', 'value', 'arg', 'env', 'force'])
+ * blueconfig.sortGetters(['default', 'value', 'arg', 'env']) // force is optional and must be the last one
  *
- * blueconfig.getGettersOrder();
+ * blueconfig.getGettersOrder()
  * // ['default', 'value', 'arg', 'env', 'force']
  *
  * @param    {string[]}    newOrder       Value of the property to validate
  *
  * @return   {this}
  */
-Blueconfig.prototype.sortGetters = function(newOrder) {
+BlueconfigCore.prototype.sortGetters = function(newOrder) {
   const sortFilter = this.Getter.sortGetters(this.Getter.storage.order, newOrder)
 
   this.Getter.storage.order.sort(sortFilter)
@@ -85,20 +110,20 @@ Blueconfig.prototype.sortGetters = function(newOrder) {
  * @example
  * convict.addGetter('aws', function(aws, schema, stopPropagation) {
  *   return aws.get(aws)
- * }, false, true);
+ * }, false, true)
  *
- * @param {String|Object}       name                      String for Getter name, `Object/Object[]` or which contains arguments:
- * @param    {String}           name.name             Getter name
- * @param    {Function}         name.getter               *See below*
- * @param    {Boolean}          [name.usedOnlyOnce=false] *See below*
- * @param    {Boolean}          [name.rewrite=false]      *See below*
+ * @param {string|object}       name                      String for Getter name, `Object/Object[]` or which contains arguments:
+ * @param    {string}           name.name             Getter name
+ * @param    {function}         name.getter               *See below*
+ * @param    {boolean}          [name.usedOnlyOnce=false] *See below*
+ * @param    {boolean}          [name.rewrite=false]      *See below*
  * @param    {ConfigObjectModel.getterCallback}   getter  Getter function get external value depending of the name name.
- * @param    {Boolean}          [usedOnlyOnce=false]      `false` by default. If true, The value can't be reused by another `keyname=value`
- * @param    {Boolean}          [rewrite=false]           Allow rewrite an existant format
+ * @param    {boolean}          [usedOnlyOnce=false]      `false` by default. If true, The value can't be reused by another `keyname=value`
+ * @param    {boolean}          [rewrite=false]           Allow rewrite an existant format
  *
  * @return   {this}
  */
-Blueconfig.prototype.addGetter = function(name, getter, usedOnlyOnce, rewrite) {
+BlueconfigCore.prototype.addGetter = function(name, getter, usedOnlyOnce, rewrite) {
   if (typeof name === 'object') {
     getter = name.getter
     usedOnlyOnce = name.usedOnlyOnce
@@ -127,17 +152,17 @@ Blueconfig.prototype.addGetter = function(name, getter, usedOnlyOnce, rewrite) {
  *     },
  *     usedOnlyOnce: true
  *   }
- * });
+ * })
  *
- * @param {Object|Object[]}    getters                      Object containing list of Getters/Object
- * @param {Object}    getters.{name}                        {name} in `getters.{name}` is the getter name
- * @param {Function}  getters.{name}.getter                 *See Blueconfig.addGetter*
- * @param {Boolean}   [getters.{name}.usedOnlyOnce=false]   *See Blueconfig.addGetter*
- * @param {Boolean}   [getters.{name}.rewrite=false]        *See Blueconfig.addGetter*
+ * @param {object|object[]}    getters                      Object containing list of Getters/Object
+ * @param {object}    getters.{name}                        {name} in `getters.{name}` is the getter name
+ * @param {function}  getters.{name}.getter                 *See Blueconfig.addGetter*
+ * @param {boolean}   [getters.{name}.usedOnlyOnce=false]   *See Blueconfig.addGetter*
+ * @param {boolean}   [getters.{name}.rewrite=false]        *See Blueconfig.addGetter*
  *
  * @return   {this}
  */
-Blueconfig.prototype.addGetters = function(getters) {
+BlueconfigCore.prototype.addGetters = function(getters) {
   if (Array.isArray(getters)) {
     getters.forEach((child) => {
       this.addGetter(child)
@@ -158,7 +183,7 @@ Blueconfig.prototype.addGetters = function(getters) {
  * Adds a new custom format. Validate function and coerce function will be used to validate COM property with `format` type.
  *
  * @example
- * BluePrint.addFormat({
+ * Blueconfig.addFormat({
  *   name: 'int',
  *   coerce: (value) => (typeof value !== 'undefined') ? parseInt(value, 10) : value,
  *   validate: function(value) {
@@ -169,18 +194,18 @@ Blueconfig.prototype.addGetters = function(getters) {
  * })
  *
  *
- * @param {String|Object}                name              String for Format name, `Object/Object[]` or which contains arguments:
- * @param    {String}                    name.name         Format name
- * @param    {Function}                  name.validate     *See below*
- * @param    {Function}                  name.coerce       *See below*
- * @param    {Boolean}               [name.rewrite=false]  *See below*
+ * @param {string|object}                name              String for Format name, `Object/Object[]` or which contains arguments:
+ * @param    {string}                    name.name         Format name
+ * @param    {function}                  name.validate     *See below*
+ * @param    {function}                  name.coerce       *See below*
+ * @param    {boolean}               [name.rewrite=false]  *See below*
  * @param {SchemaNode.validateCallback}  validate          Validate function, should throw if the value is wrong `Error` or [`LISTOFERRORS` (see example)](./ZCUSTOMERROR.LISTOFERRORS.html)
  * @param {SchemaNode.coerce}            coerce            Coerce function to convert a value to a specified function (can be omitted)
- * @param    {Boolean}                   [rewrite=false]   Allow rewrite an existant format
+ * @param    {boolean}                   [rewrite=false]   Allow rewrite an existant format
  *
  * @return   {this}
  */
-Blueconfig.prototype.addFormat = function(name, validate, coerce, rewrite) {
+BlueconfigCore.prototype.addFormat = function(name, validate, coerce, rewrite) {
   if (typeof name === 'object') {
     validate = name.validate
     coerce = name.coerce
@@ -214,15 +239,15 @@ Blueconfig.prototype.addFormat = function(name, validate, coerce, rewrite) {
  *
  * @see Blueconfig.addFormat
  *
- * @param {Object|Object[]}    formats               Object containing list of Object
- * @param {Object}    formats.{name}                 {name} in `formats.{name}` is the format name
- * @param {Function}  formats.{name}.validate        *See Blueconfig.addFormat*
- * @param {Function}  formats.{name}.coerce          *See Blueconfig.addFormat*
- * @param {Boolean}   [formats.{name}.rewrite=false] *See Blueconfig.addFormat*
+ * @param {object|object[]}    formats               Object containing list of Object
+ * @param {object}    formats.{name}                 {name} in `formats.{name}` is the format name
+ * @param {function}  formats.{name}.validate        *See Blueconfig.addFormat*
+ * @param {function}  formats.{name}.coerce          *See Blueconfig.addFormat*
+ * @param {boolean}   [formats.{name}.rewrite=false] *See Blueconfig.addFormat*
  *
  * @return   {this}
  */
-Blueconfig.prototype.addFormats = function(formats) {
+BlueconfigCore.prototype.addFormats = function(formats) {
   if (Array.isArray(formats)) {
     formats.forEach((child) => {
       this.addFormat(child)
@@ -241,20 +266,27 @@ Blueconfig.prototype.addFormats = function(formats) {
 /**
  * Adds new custom file parsers. JSON.parse will be used by default for unknown extension (default extension -> `*` => JSON).
  *
+ * Blueconfig is able to parse files with custom file types during `merge`. For this specify the
+ * corresponding parsers with the associated file extensions.
+ *
+ * If no supported extension is detected, `merge` will fallback to using the default json parser `JSON.parse`.
+ *
  * @example
  * blueconfig.addParser([
+ *  { extension: ['yml', 'yaml'], parse: yaml.safeLoad },
  *  { extension: ['yaml', 'yml'], parse: require('yaml').safeLoad },
  *  // will allow comment in json file
- *  { extension: 'json5', parse: require('json5').parse }
- * ]);
+ *  { extension: 'json', parse: require('json5').parse } // replace `JSON` by `json5`
+ * ])
+ * config.merge('config.yml')
  *
- * @param    {Object[]}    parsers              Parser
+ * @param    {object[]}    parsers              Parser
  * @param    {string}      parsers.extension    Parser extension
  * @param    {function}    parsers.parse        Parser function
  *
  * @return   {this}
  */
-Blueconfig.prototype.addParser = function(parsers) {
+BlueconfigCore.prototype.addParser = function(parsers) {
   if (!Array.isArray(parsers)) parsers = [parsers]
 
   parsers.forEach((parser) => {