+++ packages/blueconfig/lib/format/standard-formats.js
@@ -1,4 +1,4 @@
-/**
+/*
  * blueconfig-format
  * Standard format for blueconfig
  */