+++ packages/blueconfig/lib/index.js
@@ -2,16 +2,40 @@ const parseArgs = require('yargs-parser')
 
 const Core = require('./core.js')
 
-/* >> HACK TO KEEP THE SAME BEHAVIOR (= STATIC OBJECT) AND PREVENT BREAK CHANGE */
-/* SORRY IF YOU HAVE HEART ATTACK WHILE YOU READING THIS LINE ;-) I had no choice in writing this… */
-
-/*
- * @returns {Blueconfig}
+/**
+ * Static version of BlueConfigCore, with standard getters/formats:
+ *
+ * ```js
+ * Blueconfig.addGetter('default', (value, schema, stopPropagation) => schema._cvtCoerce(value))
+ * Blueconfig.sortGetters(['default', 'value']) // set default before value
+ * Blueconfig.addGetter('env', function(value, schema, stopPropagation) {
+ *   return schema._cvtCoerce(this.getEnv()[value])
+ * })
+ * Blueconfig.addGetter('arg', function(value, schema, stopPropagation) {
+ *   const argv = parseArgs(this.getArgs(), {
+ *     configuration: {
+ *       'dot-notation': false
+ *     }
+ *   })
+ *   return schema._cvtCoerce(argv[value])
+ * }, true)
+ *
+ * Blueconfig.addFormats(require('./format/standard-formats.js'))
+ * ```
+ *
+ * @example
+ * const schema = { port: { default: 80, format: 'port' } }
+ * const conf = blueconfig(schema);
+ *
+ * @class
  */
 const Blueconfig = function(rawSchema, options) {
   return Blueconfig.init(rawSchema, options)
 }
 
+
+/* >> HACK TO KEEP THE SAME BEHAVIOR (= STATIC OBJECT like convict) AND PREVENT BREAK CHANGE FOR v6.1.0 */
+/* SORRY IF YOU HAVE HEART ATTACK WHILE YOU READING THIS LINE ;-) I had no choice in writing this… */
 Core.prototype.initPerformer.call(Blueconfig)
 
 Blueconfig.init = Core.prototype.init
@@ -24,6 +48,8 @@ Blueconfig.addFormats = Core.prototype.addFormats
 Blueconfig.addParser = Core.prototype.addParser
 /* << */
 
+/* >> default getters & standard formats */
+// Add getters
 Blueconfig.addGetter('default', (value, schema, stopPropagation) => schema._cvtCoerce(value))
 Blueconfig.sortGetters(['default', 'value']) // set default before value
 Blueconfig.addGetter('env', function(value, schema, stopPropagation) {
@@ -38,7 +64,9 @@ Blueconfig.addGetter('arg', function(value, schema, stopPropagation) {
   return schema._cvtCoerce(argv[value])
 }, true)
 
+// Add formats
 Blueconfig.addFormats(require('./format/standard-formats.js'))
+/* << */
 
 module.exports = Blueconfig